+++ Makefile
@@ -74,7 +74,7 @@ update-scan-log: ## Update successful scan job pod log used in tests from templa
 	trivy image nginxinc/nginx-unprivileged@sha256:6da1811b094adbea1eb34c3e48fc2833b1a11a351ec7b36cc390e740a64fbae4 \
 		--offline-scan --severity CRITICAL,HIGH --quiet --format template \
 		--template @$(shell pwd)/internal/trivy/templates/scan-report.json.tmpl \
-		> controllers/testdata/scan-job-successful/successful-scan-job-pod.log.json
+		> internal/controller/stas/testdata/scan-job-successful/successful-scan-job-pod.log.json
 
 .PHONY: go-mod-tidy
 go-mod-tidy: ## Run go mod tidy against code.
