+++ internal/controller/stas/containerimagescan_controller.go
@@ -1,4 +1,4 @@
-package controllers
+package stas
 
 import (
 	"context"
