+++ internal/controller/stas/containerimagescan_controller_test.go
@@ -1,4 +1,4 @@
-package controllers
+package stas
 
 import (
 	"context"
