+++ internal/controller/stas/indexer.go
@@ -1,4 +1,4 @@
-package controllers
+package stas
 
 import (
 	"context"
