+++ internal/controller/stas/predicates.go
@@ -1,4 +1,4 @@
-package controllers
+package stas
 
 import (
 	"regexp"
