+++ internal/controller/stas/scan_job_controller.go
@@ -1,4 +1,4 @@
-package controllers
+package stas
 
 import (
 	"context"
