+++ internal/controller/stas/scan_job_controller_test.go
@@ -1,4 +1,4 @@
-package controllers
+package stas
 
 import (
 	"bytes"
