+++ internal/controller/stas/testdata/scan-job-failed/failed-scan-cis.yaml
