+++ internal/controller/stas/testdata/scan-job-successful-long/cis.yaml
