+++ internal/controller/stas/testdata/scan-job-successful-long/scan-job-pod.log.json
