+++ internal/controller/stas/testdata/scan-job-successful/successful-scan-cis.yaml
