+++ internal/controller/stas/testdata/scan-job-successful/successful-scan-job-pod.log.json
