+++ internal/controller/stas/testdata/scan-job/cis.yaml
