+++ internal/controller/stas/testdata/scan-job/expected-scan-job.yaml
