+++ internal/controller/stas/testdata/scan-job/workload-pod.yaml
